update dist to Xenial.
I updated dist to Xenial because trusty can only install the old chrome version and the test fails, so I updated dist to Xenial.

ref: https://travis-ci.community/t/travisci-chrome-version-has-regressed-back-to-v62/8603

It seems that the setting of ChromeDriver of Capybarah was wrong.

``` ruby
     Selenium::WebDriver::Error::UnknownError:
       unknown error: Chrome failed to start: crashed.
         (unknown error: DevToolsActivePort file doesn't exist)
         (The process started from chrome location /usr/bin/google-chrome is no longer running, so ChromeDriver is assuming that Chrome has crashed.)
```

Using Selenium::WebDriver::Remote::Capabilities was not a good idea, so I've modified it to use Selenium::WebDriver::Chrome::Options to configure it :thinking: